Advanced Website 2.0
We got rid of the the style.css file and the script.js files.
Integrated both the needed codes in the index.html file. Completely new theme and design. Deleted all the files and images from the amazon aws bucket.
And a note to keep  - clear cache everytime you make changes, both from the browser and mainly from the aws.